Adding service history, correcting migration data type (real vs numeric)

+++ back-end/migrations/2023-12-23-035111_service/up.sql
@@ -2,6 +2,6 @@ CREATE TABLE service (
 	id INTEGER PRIMARY KEY,
 	name TEXT NOT NULL,
 	description TEXT NOT NULL,
-	price NUMERIC NOT NULL,
+	price REAL NOT NULL,
 	base_time INTEGER NOT NULL,
 );
